+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* A simple approach inspired on dartz package.
+    * Left, Result and Right classes;
+    * Empty class to not return any data;